TranscriptionService/pyproject.toml
[project]
name = "transcriptionservice"
version = "0.1.0"
description = "PrepLog audio transcription microservice using OpenAI Whisper"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "numpy>=2.4.2",
    "openai-whisper>=20250625",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "redis>=7.3.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.5",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
